Create new rpcnet module, and move 'getconnectioncount' RPC to it
